# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.1.0"
  title: AA-Mock-Engine
# basePath prefixes all resource paths
basePath: /
...
schemes:
  # Uncomment for local dev:
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/octet-stream
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /flights:
    x-swagger-router-controller: flight
    get:
      operationId: flights
      tags:
          - Flight
      parameters:
        - name: origin
          in: query
          description: "Airport code for origin"
          type: string
          enum: ["DFW", "LAX", "MIA", "PHL", "ORD", "JFK", "LHR", "HKG"]
        - name: destination
          in: query
          description: "Airport code for destination"
          type: string
          enum: ["DFW", "LAX", "MIA", "PHL", "ORD", "JFK", "LHR", "HKG"]
        - name: date
          in: query
          description: "Departure date (origin) for the desired flight (e.g., \"2018-03-01\")"
          required: true
          type: string
          format: date
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            type: array
            items:
              $ref: "#/definitions/FlightSegment"
        "500":
          description: Error retrieving flights
          schema:
            type: string
  /currentFlight:
    x-swagger-router-controller: flight
    get:
      description: Returns a random flight's information for simulating current flight
      tags:
          - Flight
      operationId: currentFlight
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/FlightSegment"
        "404":
          description: Flight not found
          schema:
            type: string
        "500":
            description: Error retrieving flight
            schema:
              type: string
  /flightStatus:
    x-swagger-router-controller: flight
    post:
      operationId: setFlightStatus
      tags:
        - Flight
      parameters:
        - name: flightId
          in: query
          description: "Flight ID (MongoDB _id) for the desired flight"
          required: true
          type: string
        - name: flightStatus
          in: query
          description: "New flight status for the desired flight"
          required: true
          type: string
          default: "On-Time"
          enum: ["On-Time", "Delayed", "Cancelled"]
      responses:
        "200":
          description: Success
          schema:
            type: string
        "404":
          description: Flight not found
          schema:
            type: string
        "500":
          description: Error retrieving flight status
          schema:
            type: string
...
definitions:
...
  FlightSegment:
    properties:
      cost:
        type: number
      flightNumber:
        type: string
      originCode:
        type: string
      # originCity:
      #   type: string
      destinationCode:
        type: string
      # destinationCity:
      #   type: string
      estimatedDepartureTime:
        type: string
      scheduledDepartureTime:
        type: string
      estimatedArrivalTime:
        type: string
      scheduledArrivalTime:
        type: string
...